refector add repositories class
